[web] Update linklist and filter panel tags when a new link is created.
TODO: This could be better. As currently implemented, this causes the
entire link list and tag list to be refetched from the server and an
expensive DOM recalculation.

Instead, this should just amend the new link and any new tags to the
existing list incrementally, rather than refetching a bunch of redudant
data.